+++ Automate/Framework/GenericMachine.cs
@@ -1,5 +1,4 @@
 ﻿using StardewValley;
-using StardewValley.Objects;
 using SObject = StardewValley.Object;
 
 namespace Pathoschild.Stardew.Automate.Framework
@@ -29,24 +28,15 @@ namespace Pathoschild.Stardew.Automate.Framework
         }
 
         /// <summary>Get the output item.</summary>
-        /// <remarks>This should have no effect on the machine state, since the chests may not have room for the item.</remarks>
-        public virtual Item GetOutput()
+        public virtual ITrackedStack GetOutput()
         {
-            return this.Machine.heldObject;
+            return new TrackedItem(this.Machine.heldObject, onEmpty: this.GenericReset);
         }
 
-        /// <summary>Reset the machine so it's ready to accept a new input.</summary>
-        /// <param name="outputTaken">Whether the current output was taken.</param>
-        public virtual void Reset(bool outputTaken)
-        {
-            this.Machine.heldObject = null;
-            this.Machine.readyForHarvest = false;
-        }
-
-        /// <summary>Pull items from the connected chests.</summary>
-        /// <param name="chests">The connected chests.</param>
+        /// <summary>Pull items from the connected pipes.</summary>
+        /// <param name="pipes">The connected IO pipes.</param>
         /// <returns>Returns whether the machine started processing an item.</returns>
-        public abstract bool Pull(Chest[] chests);
+        public abstract bool Pull(IPipe[] pipes);
 
 
         /*********
@@ -57,6 +47,14 @@ namespace Pathoschild.Stardew.Automate.Framework
         protected GenericMachine(TMachine machine)
         {
             this.Machine = machine;
+        }
+
+        /// <summary>Reset the machine so it's ready to accept a new input.</summary>
+        /// <param name="item">The output item that was taken.</param>
+        protected void GenericReset(Item item)
+        {
+            this.Machine.heldObject = null;
+            this.Machine.readyForHarvest = false;
         }
     }