+++ Automate/Framework/IMachine.cs
@@ -1,10 +1,7 @@
-﻿using StardewValley;
-using StardewValley.Objects;
-
-namespace Pathoschild.Stardew.Automate.Framework
+﻿namespace Pathoschild.Stardew.Automate.Framework
 {
     /// <summary>A machine that accepts input and provides output.</summary>
-    public interface IMachine
+    internal interface IMachine
     {
         /*********
         ** Public methods
@@ -13,16 +10,11 @@ namespace Pathoschild.Stardew.Automate.Framework
         MachineState GetState();
 
         /// <summary>Get the output item.</summary>
-        /// <remarks>This should have no effect on the machine state, since the chests may not have room for the item.</remarks>
-        Item GetOutput();
-
-        /// <summary>Reset the machine so it's ready to accept a new input.</summary>
-        /// <param name="outputTaken">Whether the current output was taken.</param>
-        void Reset(bool outputTaken);
+        ITrackedStack GetOutput();
 
         /// <summary>Pull items from the connected chests.</summary>
-        /// <param name="chests">The connected chests.</param>
+        /// <param name="pipes">The connected IO pipes.</param>
         /// <returns>Returns whether the machine started processing an item.</returns>
-        bool Pull(Chest[] chests);
+        bool Pull(IPipe[] pipes);
     }
 }