+++ Automate/Machines/Objects/MushroomBoxMachine.cs
@@ -1,5 +1,4 @@
 ﻿using Pathoschild.Stardew.Automate.Framework;
-using StardewValley.Objects;
 using SObject = StardewValley.Object;
 
 namespace Pathoschild.Stardew.Automate.Machines.Objects
@@ -23,10 +22,10 @@ namespace Pathoschild.Stardew.Automate.Machines.Objects
                 : MachineState.Processing;
         }
 
-        /// <summary>Pull items from the connected chests.</summary>
-        /// <param name="chests">The connected chests.</param>
+        /// <summary>Pull items from the connected pipes.</summary>
+        /// <param name="pipes">The connected IO pipes.</param>
         /// <returns>Returns whether the machine started processing an item.</returns>
-        public override bool Pull(Chest[] chests)
+        public override bool Pull(IPipe[] pipes)
         {
             return false; // no input needed
         }