+++ Automate/Machines/Objects/SeedMakerMachine.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using Pathoschild.Stardew.Automate.Framework;
 using StardewValley;
-using StardewValley.Objects;
 using SObject = StardewValley.Object;
 
 namespace Pathoschild.Stardew.Automate.Machines.Objects
@@ -25,18 +24,18 @@ namespace Pathoschild.Stardew.Automate.Machines.Objects
         public SeedMakerMachine(SObject machine)
             : base(machine) { }
 
-        /// <summary>Pull items from the connected chests.</summary>
-        /// <param name="chests">The connected chests.</param>
+        /// <summary>Pull items from the connected pipes.</summary>
+        /// <param name="pipes">The connected IO pipes.</param>
         /// <returns>Returns whether the machine started processing an item.</returns>
-        public override bool Pull(Chest[] chests)
+        public override bool Pull(IPipe[] pipes)
         {
             SObject machine = this.Machine;
 
             // crop => seeds
-            if (chests.TryGetIngredient(this.IsValidCrop, 1, out Requirement crop))
+            if (pipes.TryGetIngredient(this.IsValidCrop, 1, out Requirement crop))
             {
-                crop.Consume();
-                int seedID = SeedMakerMachine.CropSeedIDs[crop.GetOne().parentSheetIndex];
+                crop.Reduce();
+                int seedID = SeedMakerMachine.CropSeedIDs[crop.Sample.parentSheetIndex];
 
                 Random random = new Random((int)Game1.stats.DaysPlayed + (int)Game1.uniqueIDForThisGame / 2 + (int)machine.tileLocation.X + (int)machine.tileLocation.Y * 77 + Game1.timeOfDay);
                 machine.heldObject = new SObject(seedID, random.Next(1, 4));
@@ -57,11 +56,11 @@ namespace Pathoschild.Stardew.Automate.Machines.Objects
         *********/
         /// <summary>Get whether a given item is a crop compatible with the seed marker.</summary>
         /// <param name="item">The item to check.</param>
-        public bool IsValidCrop(Item item)
+        public bool IsValidCrop(ITrackedStack item)
         {
             return
-                item.parentSheetIndex != 433 // seed maker doesn't allow coffee beans
-                && SeedMakerMachine.CropSeedIDs.ContainsKey(item.parentSheetIndex);
+                item.Sample.parentSheetIndex != 433 // seed maker doesn't allow coffee beans
+                && SeedMakerMachine.CropSeedIDs.ContainsKey(item.Sample.parentSheetIndex);
         }
 
         /// <summary>Get a crop ID => seed ID lookup.</summary>
@@ -74,7 +73,8 @@ namespace Pathoschild.Stardew.Automate.Machines.Objects
             {
                 int dataCropID = Convert.ToInt32(entry.Value.Split('/')[3]);
                 int dataSeedID = entry.Key;
-                lookup.Add(dataCropID, dataSeedID);
+                if (!lookup.ContainsKey(dataCropID)) // if multiple crops have the same seed, use the earliest one (which is more likely the vanilla seed)
+                    lookup[dataCropID] = dataSeedID;
             }
 
             return lookup;