+++ Automate/Machines/Objects/TapperMachine.cs
@@ -1,7 +1,6 @@
 ﻿using Microsoft.Xna.Framework;
 using Pathoschild.Stardew.Automate.Framework;
 using StardewValley;
-using StardewValley.Objects;
 using StardewValley.TerrainFeatures;
 using SObject = StardewValley.Object;
 
@@ -36,15 +35,26 @@ namespace Pathoschild.Stardew.Automate.Machines.Objects
         }
 
         /// <summary>Get the output item.</summary>
-        /// <remarks>This should have no effect on the machine state, since the chests may not have room for the item.</remarks>
-        public override Item GetOutput()
+        public override ITrackedStack GetOutput()
         {
-            return this.Machine.heldObject.getOne();
+            return new TrackedItem(this.Machine.heldObject.getOne(), onEmpty: this.Reset);
         }
 
+        /// <summary>Pull items from the connected pipes.</summary>
+        /// <param name="pipes">The connected IO pipes.</param>
+        /// <returns>Returns whether the machine started processing an item.</returns>
+        public override bool Pull(IPipe[] pipes)
+        {
+            return false; // no input
+        }
+
+
+        /*********
+        ** Private methods
+        *********/
         /// <summary>Reset the machine so it's ready to accept a new input.</summary>
-        /// <param name="outputTaken">Whether the current output was taken.</param>
-        public override void Reset(bool outputTaken)
+        /// <param name="item">The output item that was taken.</param>
+        private void Reset(Item item)
         {
             SObject tapper = this.Machine;
 
@@ -74,14 +84,6 @@ namespace Pathoschild.Stardew.Automate.Machines.Objects
             }
             tapper.heldObject = (SObject)tapper.heldObject.getOne();
             tapper.readyForHarvest = false;
-        }
-
-        /// <summary>Pull items from the connected chests.</summary>
-        /// <param name="chests">The connected chests.</param>
-        /// <returns>Returns whether the machine started processing an item.</returns>
-        public override bool Pull(Chest[] chests)
-        {
-            return false; // no input
         }
     }
 }