+++ Automate/Machines/TerrainFeatures/FruitTreeMachine.cs
@@ -1,6 +1,5 @@
 ﻿using Pathoschild.Stardew.Automate.Framework;
 using StardewValley;
-using StardewValley.Objects;
 using StardewValley.TerrainFeatures;
 
 namespace Pathoschild.Stardew.Automate.Machines.TerrainFeatures
@@ -37,14 +36,13 @@ namespace Pathoschild.Stardew.Automate.Machines.TerrainFeatures
         }
 
         /// <summary>Get the output item.</summary>
-        /// <remarks>This should have no effect on the machine state, since the chests may not have room for the item.</remarks>
-        public Item GetOutput()
+        public ITrackedStack GetOutput()
         {
             FruitTree tree = this.Tree;
 
             // if struck by lightning => coal
             if (tree.struckByLightningCountdown > 0)
-                return new Object(382, tree.fruitsOnTree);
+                return new TrackedItem(new Object(382, tree.fruitsOnTree), onReduced: this.OnOutputReduced);
 
             // else => fruit
             int quality = 0;
@@ -54,22 +52,26 @@ namespace Pathoschild.Stardew.Automate.Machines.TerrainFeatures
                 quality = 2;
             else if (tree.daysUntilMature <= -336)
                 quality = 4;
-            return new Object(tree.indexOfFruit, tree.fruitsOnTree, quality: quality);
+            return new TrackedItem(new Object(tree.indexOfFruit, tree.fruitsOnTree, quality: quality), onReduced: this.OnOutputReduced);
         }
 
-        /// <summary>Reset the machine so it's ready to accept a new input.</summary>
-        /// <param name="outputTaken">Whether the current output was taken.</param>
-        public void Reset(bool outputTaken)
-        {
-            this.Tree.fruitsOnTree = 0;
-        }
-
-        /// <summary>Pull items from the connected chests.</summary>
-        /// <param name="chests">The connected chests.</param>
+        /// <summary>Pull items from the connected pipes.</summary>
+        /// <param name="pipes">The connected IO pipes.</param>
         /// <returns>Returns whether the machine started processing an item.</returns>
-        public bool Pull(Chest[] chests)
+        public bool Pull(IPipe[] pipes)
         {
             return false; // no input
+        }
+
+
+        /*********
+        ** Private methods
+        *********/
+        /// <summary>Reset the machine so it's ready to accept a new input.</summary>
+        /// <param name="item">The output item that was taken.</param>
+        private void OnOutputReduced(Item item)
+        {
+            this.Tree.fruitsOnTree = item.Stack;
         }
     }
 }