+++ Automate/Properties/AssemblyInfo.cs
@@ -3,5 +3,5 @@ using System.Runtime.InteropServices;
 
 [assembly: AssemblyTitle("Automate")]
 [assembly: Guid("a72152b0-ab75-44e4-86d8-f6b8b9920b59")]
-[assembly: AssemblyVersion("1.1.0.0")]
-[assembly: AssemblyFileVersion("1.1.0.0")]
+[assembly: AssemblyVersion("1.2.0.0")]
+[assembly: AssemblyFileVersion("1.2.0.0")]