+++ FastAnimations/release-notes.md
@@ -1,6 +1,9 @@
 ﻿[← back to readme](README.md)
 
 # Release notes
+## 1.1
+* Fixed errors with drinking animation.
+
 ## 1.0
 * Initial version.
 * Added support for eating, drinking, milking, shearing, and geode breaking.